add exception handling in PostService

+++ src/main/java/com/gustcustodio/blogging_platform_api/services/PostService.java
@@ -4,7 +4,12 @@ import com.gustcustodio.blogging_platform_api.dtos.PostDTO;
 import com.gustcustodio.blogging_platform_api.entities.Post;
 import com.gustcustodio.blogging_platform_api.mappers.PostMapper;
 import com.gustcustodio.blogging_platform_api.repositories.PostRepository;
+import com.gustcustodio.blogging_platform_api.services.exceptions.DatabaseException;
+import com.gustcustodio.blogging_platform_api.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -23,9 +28,8 @@ public class PostService {
 
     @Transactional(readOnly = true)
     public PostDTO getSinglePost(Long id) {
-        Post post = postRepository.findById(id).orElseThrow();
-        PostDTO postDTO = postMapper.convertEntityToDto(post);
-        return postDTO;
+        Post post = postRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
+        return postMapper.convertEntityToDto(post);
     }
 
     @Transactional(readOnly = true)
@@ -43,15 +47,26 @@ public class PostService {
 
     @Transactional
     public PostDTO updatePost(Long id, PostDTO postDTO) {
-        Post post = postRepository.getReferenceById(id);
-        postMapper.updateEntityFromDto(postDTO, post);
-        post = postRepository.save(post);
-        return postMapper.convertEntityToDto(post);
+        try {
+            Post post = postRepository.getReferenceById(id);
+            postMapper.updateEntityFromDto(postDTO, post);
+            post = postRepository.save(post);
+            return postMapper.convertEntityToDto(post);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void deletePost(Long id) {
-        postRepository.deleteById(id);
+        if (!postRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+        try {
+            postRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Referential integrity failure");
+        }
     }
 
 }
